+++ mobile/android/android-components/components/browser/menu/src/main/java/mozilla/components/browser/menu/view/DynamicWidthRecyclerView.kt
@@ -19,6 +19,11 @@ class DynamicWidthRecyclerView @JvmOverloads constructor(
     context: Context,
     attrs: AttributeSet? = null
 ) : RecyclerView(context, attrs) {
+    @VisibleForTesting
+    @Px internal var maxWidthOfAllChildren: Int = 0
+        set(value) {
+            if (field == 0) field = value
+        }
 
     @Px var minWidth: Int = -1
     @Px var maxWidth: Int = -1
@@ -29,8 +34,13 @@ class DynamicWidthRecyclerView @JvmOverloads constructor(
             // Ignore any bounds set in xml. Allow for children to expand entirely.
             callParentOnMeasure(MeasureSpec.makeMeasureSpec(0, MeasureSpec.UNSPECIFIED), heightSpec)
 
+            // First measure will report the width/height for the entire list
+            // The first layout pass will actually remove child views that do not fit the screen
+            // so future onMeasure calls will report skewed values.
+            maxWidthOfAllChildren = measuredWidth
+
             // Children now have "unspecified" width. Let's set some bounds.
-            setReconciledDimensions(measuredWidth, measuredHeight)
+            setReconciledDimensions(maxWidthOfAllChildren, measuredHeight)
         } else {
             // Default behavior. layout_width / layout_height properties will be used for measuring.
             callParentOnMeasure(widthSpec, heightSpec)